2-strncpy.c

+++ 0x06-pointers_arrays_strings/2-strncpy.c
@@ -1,5 +1,5 @@
-#include <stdio.h>
 #include <string.h>
+#include "main.h"
 
 /**
  * _strncpy - string copy
@@ -11,7 +11,7 @@
 
 char *_strncpy(char *dest, char *src, int n)
 {
-	size_t i;
+	int i;
 
 	for (i = 0; i < n && src[i] != '\0'; i++)
 	{
